Implement quantized power functions, edited testbenches.

+++ src/StatisticsFunctions/plp_power_i8.c
@@ -72,7 +72,7 @@
 void plp_power_i8(
                          const int8_t * __restrict__ pSrc,
                          uint32_t blockSize,
-                         int8_t * __restrict__ pRes){
+                         int32_t * __restrict__ pRes){
   
   if (rt_cluster_id() == ARCHI_FC_CID){
     plp_power_i8s_rv32im(pSrc, blockSize, pRes);
